requirements.txt
Django==4.2.16
djangorestframework==3.15.2
python-decouple==3.8
mysql-connector-python==8.3.0
pymysql==1.1.1
paramiko==3.4.0
netaddr==1.3.0
requests==2.32.3
